Fix systemd-tmpfiles-setup.service in initrd.img
To reduce the size of initrd, most files in /usr/share/ are deleted.
However, systemd-tmpfiles-setup.service failed to start because it
requires in /usr/share/factory/etc/nsswitch.conf and
/usr/share/factory/etc/pam.d.
Add /usr/share/factory to fix systemd-tmpfiles-setup.service

+++ installer/mk-install-iso.sh
@@ -257,7 +257,7 @@ if [ "$LIVE_CD" = false ] ; then
 
     # TODO: mbassiouny, Find a clean way to do that
     for i in `ls ${BUILDROOT}/usr/share/`; do
-    	if [ $i != 'terminfo' -a $i != 'cracklib' -a $i != 'grub' ]; then
+    	if [ $i != 'terminfo' -a $i != 'cracklib' -a $i != 'grub' -a $i != 'factory' ]; then
     		rm -rf ${BUILDROOT}/usr/share/$i
     	fi
     done
